backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12
tensorflow==2.21.0
pillow==10.4.0
numpy==2.2.1
scikit-learn==1.7.2
xgboost==2.1.4
joblib==1.4.2
pandas==2.2.3
httpx==0.27.2
requests==2.32.3
opencv-python-headless==4.10.0.84
boto3==1.35.74
azure-storage-blob==12.23.1
pydantic==2.9.2
python-dotenv==1.0.1
rasterio==1.4.2
earthaccess==0.10.0